12/04 :recycle: : getUserId 추출 후 중복 코드 리팩토링
