+++ src/app/api/tags/[id]/route.ts
@@ -1,9 +1,8 @@
 import ERROR_MESSAGES from '@/app/lib/constants/errorMessages';
-import { verifyAuthToken } from '@/app/services/token/verifyToken';
 import { NextResponse } from 'next/server';
-import jwt from 'jsonwebtoken';
 import Tag from '@/app/lib/db/models/tags';
 import dbConnect from '@/app/lib/db/dbConnect';
+import { getUserId } from '@/app/services/token/getUserId';
 
 export async function GET(
   req: Request,
@@ -12,20 +11,9 @@ export async function GET(
   try {
     await dbConnect();
 
-    const { error, token } = verifyAuthToken(req);
-    if (error) {
-      return NextResponse.json({ error }, { status: 401 });
-    }
-
-    const JWT_SECRET = process.env.JWT_SECRET;
-    if (!JWT_SECRET)
-      return NextResponse.json(
-        { error: ERROR_MESSAGES.JWT_SECRET_ERROR.ko },
-        { status: 500 }
-      );
-
-    const decoded = jwt.verify(token!, JWT_SECRET) as { id: string };
-    const loggedInUserId = decoded.id;
+    const { userId, error } = getUserId(req);
+    if (!userId) return NextResponse.json({ error }, { status: 401 });
+    const loggedInUserId = userId;
     const requestedUserId = params.id;
 
     if (loggedInUserId === requestedUserId) {
