+++ src/app/api/tags/route.ts
@@ -3,27 +3,8 @@ import { NextRequest, NextResponse } from 'next/server';
 import Tag from '@/app/lib/db/models/tags';
 import { updateTagAndChecks } from '@/app/services/database/updateTagAndChecks';
 import { deleteTagAndChecks } from '@/app/services/database/deleteTagAndChecks';
-import { verifyAuthToken } from '@/app/services/token/verifyToken';
 import ERROR_MESSAGES from '@/app/lib/constants/errorMessages';
-import jwt from 'jsonwebtoken';
-
-const getUserId = (req: Request): { userId: string; error?: string } => {
-  let userId = '';
-
-  const { error, token } = verifyAuthToken(req);
-
-  if (error) {
-    return { userId, error };
-  }
-
-  const JWT_SECRET = process.env.JWT_SECRET;
-  if (!JWT_SECRET) return { userId, error: ERROR_MESSAGES.JWT_SECRET_ERROR.ko };
-  const decoded = jwt.verify(token!, JWT_SECRET) as { id: string };
-
-  userId = decoded.id;
-
-  return { userId };
-};
+import { getUserId } from '@/app/services/token/getUserId';
 
 export async function GET() {
   try {
